validateShippingTime in Setup MC shipping time.
